Re-arranged settings naming scheme
Settings which are not on a per-file basis are now "Globals".
Some settings have been moved around - the tangle now has a per-file
Settings struct as well as the Globals struct.
Css is now open to move back to Settings instead of being Globals - I'll
eventually do this for the sake of correctness, but it will still be
warned against.
`out_dir` has been moved from the top-level Globals to being a field on
each of the tangle and weave Globals structs - this is to allow for the
directories to be different in the future.