Eliminated duplication between parser/linker
LitFile and LinkedLitFile had a bunch of copied fields, since the design
was put together to work as quickly as possible.
In retrospect, this was both bad and added a lot of debt to expanding
the system.
Instead, file metadata is now stored in a separate struct - dereffing is
used to minimise code problems - and that struct is shared between the
parsing/linking data structures.